core/requirements.txt
flask
pytest
python-dotenv
requests
pathspec
responses
pytest-cov>=2.12.0,<3.0.0
